backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
paho-mqtt==1.6.1
pydantic-settings==2.2.1
websockets==12.0
sqlalchemy==2.0.27
aiosqlite==0.19.0
python-dateutil==2.8.2
